requirements.txt
flask
pymupdf
opencv-python
numpy
Pillow
reportlab